requirements.txt
requests
bs4
html5lib
numpy
python-dotenv
fastapi-caching